requirements.txt
google-cloud-aiplatform
google-cloud-bigquery
requests
streamlit